Table: Functionally GA the refactor by ignoring flag
