+++ packages/grafana-ui/src/components/Table/TableNG/TableNested.tsx
@@ -13,13 +13,15 @@ import {
   type SortColumn,
 } from '@grafana/react-data-grid';
 
-import { useStyles2, useTheme2 } from '../../../../themes/ThemeContext';
-import { getTextColorForBackground as _getTextColorForBackground } from '../../../../utils/colors';
-import { usePanelContext } from '../../../PanelChrome';
-import { type DataLinksActionsTooltipState } from '../../cellUtils';
-import { EmptyTablePlaceholder } from '../components/EmptyTablePlaceholder';
-import { RowExpander } from '../components/RowExpander';
-import { COLUMN, TABLE } from '../constants';
+import { useStyles2, useTheme2 } from '../../../themes/ThemeContext';
+import { getTextColorForBackground as _getTextColorForBackground } from '../../../utils/colors';
+import { usePanelContext } from '../../PanelChrome';
+import { type DataLinksActionsTooltipState } from '../cellUtils';
+
+import { TableDataGrid } from './TableDataGrid';
+import { EmptyTablePlaceholder } from './components/EmptyTablePlaceholder';
+import { RowExpander } from './components/RowExpander';
+import { COLUMN, TABLE } from './constants';
 import {
   useColumnResize,
   useColWidths,
@@ -33,8 +35,9 @@ import {
   useRowHeight,
   useScrollbarWidth,
   useSortedRows,
-} from '../hooks';
-import { getGridStyles, IS_SAFARI_26 } from '../styles';
+} from './hooks';
+import { type ColumnBuildConfig, useColumnBuilderFromFields, useDataGridRows } from './render-hooks';
+import { getGridStyles, IS_SAFARI_26 } from './styles';
 import {
   type CellRootRenderer,
   type FromFieldsResult,
@@ -43,7 +46,7 @@ import {
   type TableNGProps,
   type TableRow,
   type TableSummaryRow,
-} from '../types';
+} from './types';
 import {
   calculateFooterHeight,
   createTypographyContext,
@@ -55,10 +58,7 @@ import {
   getDisplayName,
   getStableRowKey,
   getVisibleFields,
-} from '../utils';
-
-import { TableDataGrid } from './TableDataGrid';
-import { type ColumnBuildConfig, useColumnBuilderFromFields, useDataGridRows } from './render-hooks';
+} from './utils';
 
 const EXPANDED_COLUMN_KEY = 'expanded';
 type OnCellClick = NonNullable<DataGridProps<TableRow, TableSummaryRow>['onCellClick']>;
