+++ packages/grafana-ui/src/components/Table/TableNG/render-hooks.tsx
@@ -37,12 +37,13 @@ import {
   type TableFooterOptions,
 } from '@grafana/schema';
 
-import { type PanelContext } from '../../../PanelChrome';
-import { getCellRenderer, getCellSpecificStyles } from '../Cells/renderers';
-import { HeaderCell } from '../components/HeaderCell';
-import { SummaryCell } from '../components/SummaryCell';
-import { TableCellActions } from '../components/TableCellActions';
-import { TableCellTooltip } from '../components/TableCellTooltip';
+import { type PanelContext } from '../../PanelChrome';
+
+import { getCellRenderer, getCellSpecificStyles } from './Cells/renderers';
+import { HeaderCell } from './components/HeaderCell';
+import { SummaryCell } from './components/SummaryCell';
+import { TableCellActions } from './components/TableCellActions';
+import { TableCellTooltip } from './components/TableCellTooltip';
 import {
   getCellActionStyles,
   getDefaultCellStyles,
@@ -52,7 +53,7 @@ import {
   getTooltipStyles,
   getJustifyContent,
   IS_SAFARI_26,
-} from '../styles';
+} from './styles';
 import {
   type CellRootRenderer,
   type FilterType,
@@ -64,7 +65,7 @@ import {
   type TableFilterActionCallback,
   type TableRow,
   type TableSummaryRow,
-} from '../types';
+} from './types';
 import {
   type ApplyFilterResult,
   canFieldBeColorized,
@@ -79,7 +80,7 @@ import {
   predicateByName,
   shouldTextOverflow,
   shouldTextWrap,
-} from '../utils';
+} from './utils';
 
 // -----------------------------------------------------------------------------
 // useDataGridRows
