+++ packages/grafana-ui/src/components/Table/TableNG/styles.ts
@@ -191,7 +191,6 @@ export const getGridStyles = memoize((theme: GrafanaTheme2, enablePagination?: b
       padding: theme.spacing(0, 1, 0, 2),
     }),
     menuItem: css({ maxWidth: '200px' }),
-    safariWrapper: css({ contain: 'strict', height: '100%' }),
   };
 });
