+++ public/app/features/table/hooks.ts
@@ -10,7 +10,6 @@ import {
   type InterpolateFunction,
 } from '@grafana/data';
 import { config } from '@grafana/runtime';
-import { useFlagTableRefactorNested } from '@grafana/runtime/internal';
 import { type TableOptions } from '@grafana/schema';
 import { usePanelContext } from '@grafana/ui';
 import { getConfig } from 'app/core/config';
@@ -74,8 +73,6 @@ type CommonTableOptions = Pick<
  * are left to the caller. Spread the result onto `<TableNG {...props} />`.
  */
 export function useCommonTableProps(options: CommonTableOptions, fieldConfig: FieldConfigSource) {
-  const nestedRefactorEnabled = useFlagTableRefactorNested();
-
   return useMemo(
     () => ({
       noHeader: !options.showHeader,
@@ -89,7 +86,6 @@ export function useCommonTableProps(options: CommonTableOptions, fieldConfig: Fi
       maxRowHeight: options.maxRowHeight,
       disableKeyboardEvents: options.disableKeyboardEvents,
       disableSanitizeHtml: getConfig().disableSanitizeHtml,
-      nestedRefactorEnabled,
     }),
     [
       options.showHeader,
@@ -101,7 +97,6 @@ export function useCommonTableProps(options: CommonTableOptions, fieldConfig: Fi
       options.maxRowHeight,
       options.disableKeyboardEvents,
       fieldConfig.defaults.noValue,
-      nestedRefactorEnabled,
     ]
   );
 }
